Added attribute checks, moved telemetry methods to separate file.
