+++ model-optimizer/extensions/load/mxnet/loader.py
@@ -17,7 +17,7 @@ from mo.front.mxnet.extractor import mxnet_op_extractors, mxnet_op_extractor
 from mo.front.mxnet.loader import symbol2nx, load_symbol_def
 from mo.front.mxnet.nd_to_params import save_params_file
 from mo.graph.graph import Graph
-from mo.utils.graph import send_shapes_info, send_op_names_info
+from mo.utils.telemetry_utils import send_shapes_info, send_op_names_info
 
 
 class MxNetLoader(Loader):
