+++ model-optimizer/mo/front/tf/loader.py
@@ -261,9 +261,13 @@ def protobuf_attrs(pb:tf_v1.NodeDef):
 def protobuf2nx(graph, pb: tf_v1.GraphDef):
     fill_graph_with_nodes(graph, pb.node, get_id=lambda pb: pb.name, get_attrs=protobuf_attrs)
 
-    for node_name in graph.nodes:
-        node = Node(graph, node_name)
-        graph.op_names_statistic[node.pb.op] += 1
+    if hasattr(graph, 'op_names_statistic'):
+        for node_name in graph.nodes:
+            node = Node(graph, node_name)
+            node_pb = node.soft_get('pb', None)
+            if node_pb is not None:
+                if hasattr(node_pb, 'op'):
+                    graph.op_names_statistic[node_pb.op] += 1
 
     # Create a library with auxiliary functions used in TensorFlow 2 operations
     if hasattr(pb, 'library') and hasattr(pb.library, 'function'):
