+++ model-optimizer/mo/main.py
@@ -36,7 +36,8 @@ from mo.utils.get_ov_update_message import get_ov_update_message
 from mo.utils.guess_framework import deduce_framework_by_namespace
 from mo.utils.logger import init_logger
 from mo.utils.model_analysis import AnalysisResults
-from mo.utils.utils import refer_to_faq_msg, send_params_info, send_framework_info
+from mo.utils.utils import refer_to_faq_msg
+from mo.utils.telemetry_utils import send_params_info, send_framework_info
 from mo.utils.version import get_version, get_simplified_mo_version, get_simplified_ie_version
 from mo.utils.versions_checker import check_requirements  # pylint: disable=no-name-in-module
