+++ model-optimizer/mo/utils/cli_parser.py
@@ -416,6 +416,15 @@ def get_onnx_cli_options():
     return OrderedDict(sorted(d.items(), key=lambda t: t[0]))
 
 
+def get_params_with_paths_list():
+    return ['input_model', 'output_dir', 'caffe_parser_path', 'extensions', 'k', 'output_dir',
+            'input_checkpoint', 'input_meta_graph', 'input_proto', 'input_symbol', 'mean_file',
+            'mean_file_offsets', 'pretrained_model_name', 'saved_model_dir', 'tensorboard_logdir',
+            'tensorflow_custom_layer_libraries', 'tensorflow_custom_operations_config_update',
+            'tensorflow_object_detection_api_pipeline_config', 'tensorflow_use_custom_operations_config',
+            'transformations_config']
+
+
 def get_caffe_cli_parser(parser: argparse.ArgumentParser = None):
     """
     Specifies cli arguments for Model Optimizer for Caffe*
